pipecat-pipeline/requirements.txt
pandas
pipecat-ai
transformers
pytorch-lightning
tf-keras
